refactoring code duplication
