+++ Objekte-Plugin/src/main/java/entities/SecurityGate.java
@@ -1,12 +1,9 @@
 package entities;
 
 import Plugin.Task.Task;
-import Plugin.Task.TaskType;
 import dhbw.sose2022.softwareengineering.airportagentsim.simulation.api.geometry.Point;
 import dhbw.sose2022.softwareengineering.airportagentsim.simulation.api.plugin.Plugin;
-import dhbw.sose2022.softwareengineering.airportagentsim.simulation.api.simulation.World;
 import dhbw.sose2022.softwareengineering.airportagentsim.simulation.api.simulation.entity.Entity;
-import dhbw.sose2022.softwareengineering.airportagentsim.simulation.api.simulation.entity.StaticEntity;
 import dhbw.sose2022.softwareengineering.airportagentsim.simulation.api.simulation.message.LocalMessage;
 import dhbw.sose2022.softwareengineering.airportagentsim.simulation.api.simulation.message.Message;
 import dhbw.sose2022.softwareengineering.airportagentsim.simulation.simulation.SimulationWorld;
@@ -16,33 +13,14 @@ import java.util.function.Consumer;
 
 public class SecurityGate extends Door {
 
-    ArrayList<Task> tasks;
+    TaskConnector taskConnector;
     public SecurityGate(int length, int height, int xPos, int yPos, SimulationWorld world, Plugin plugin, boolean isOpen){
         super(length, height, xPos, yPos, world, plugin, isOpen);
     }
 
     public void pluginUpdate() {
         if (this.getWorld().getIteration() == 0) {
-            tasks = new ArrayList<Task>();
-            this.getWorld().getEntities().forEach(
-                    new Consumer<Entity>() {
-                        @Override
-                        public void accept(Entity entity) {
-                            if (entity instanceof Task) {
-                                if (((Task) entity).taskIsApplicable(TaskType.PERFORM_SECURITY_CHECK)) {
-                                    tasks.add((Task) entity);
-                                }
-                            }
-                        }
-                    }
-            );
-
-            tasks.forEach(new Consumer<Task>() {
-                @Override
-                public void accept(Task task) {
-                    getWorld().sendMessage(new SecurityGateCheck(xPos, yPos));
-                }
-            });
+            taskConnector.connectTasks();
         }
     }
