+++ Objekte-Plugin/src/main/java/plugin/ObjectPlugin.java
@@ -1,11 +1,18 @@
 package plugin;
 
+import Plugin.Task.Task;
+import Plugin.Task.TaskType;
 import dhbw.sose2022.softwareengineering.airportagentsim.simulation.api.AirportAgentSimulation;
 import dhbw.sose2022.softwareengineering.airportagentsim.simulation.api.config.ConfigurationFormatException;
+import dhbw.sose2022.softwareengineering.airportagentsim.simulation.api.simulation.entity.Entity;
 import entities.BaggageDropOff;
 import entities.*;
 
+import java.util.ArrayList;
+import java.util.function.Consumer;
+
 public class ObjectPlugin implements dhbw.sose2022.softwareengineering.airportagentsim.simulation.api.plugin.Plugin {
+
 
     @Override
     public void activate() {
@@ -20,4 +27,5 @@ public class ObjectPlugin implements dhbw.sose2022.softwareengineering.airportag
         }
 
     }
+
 }
